requirements.txt
requests
pydantic
fastapi
uvicorn
pycryptodome
lxml
httpx
python-dotenv
scalar-fastapi
